+++ pfscf/content/ListStore_test.go
@@ -8,7 +8,7 @@ import (
 
 func TestStore_InheritFrom(t *testing.T) {
 	t.Run("ensure deep copies", func(t *testing.T) {
-		var tc textCell
+		var tc text
 		tc.X = 1.0
 		tc.Presets = []string{"foo", "bar"}
 
@@ -18,8 +18,8 @@ func TestStore_InheritFrom(t *testing.T) {
 		s2 := NewListStore()
 		s2.InheritFrom(s1)
 
-		tcStore1 := s1[0].(*textCell)
-		tcStore2 := s2[0].(*textCell)
+		tcStore1 := s1[0].(*text)
+		tcStore2 := s2[0].(*text)
 
 		// modifications in an entry from one store should not be reflected in the other
 		tcStore1.X = 2.0